backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
openai==0.28.0
pydantic-settings==2.0.3
